リファクタリング: マウスアクションのdip_x/dip_y削除・ShowToast/Copy統一・ThemeConstants改善
PR #84 のリファクタリングでスコープ外としてスキップした3つの残課題を対処。

- マウスアクション構造体8つから未使用の dip_x/dip_y フィールドを削除
- コンテキストメニューの IDM_COPY を Dispatch(CopyClipboardAction{}) 経由に統一
- App::ShowToast を SideEffectExecutor 経由に変更し副作用の実行経路を一本化
- SideEffectExecutor に ExecuteOne() を追加し単一 SideEffect の実行を可能に
- ThemeConstants を独立ヘッダーに切り出し Theme::ToReducerConstants() を追加
