+++ src/theme/theme.cpp
@@ -18,6 +18,19 @@ float Theme::GetHeadingSize(int level) const noexcept
     return font_size_h[level - 1];
 }
 
+ThemeConstants Theme::ToReducerConstants() const noexcept
+{
+    return ThemeConstants{
+        .pane_item_height = pane_item_height,
+        .pane_header_height = pane_header_height,
+        .splitter_width = splitter_width,
+        .margin_left = margin_left,
+        .margin_right = margin_right,
+        .heading_spacing_above = heading_spacing_above,
+        .zoom = zoom,
+    };
+}
+
 void Theme::ApplyZoom(float new_zoom) noexcept
 {
     if (new_zoom <= 0.0f || zoom <= 0.0f) {
